requirements.txt
python-dotenv
pillow
flask
flask-cors
google-generativeai
DrissionPage
cryptography

# Dev / testing
pytest
pytest-cov